Replace dcterms:type with dc:type, deprecate dcterms:type
Move dcterms:type below the recommended terms and change status to deprecated. Change its notes to indicate that dc:type should be used for strings and that the RDF guide recommends rdf:type instead of dcterms:type when providing an IRI value. Replace dcterms:type in the record-level terms with dc:type and carry over the examples that were previously in dcterms:type to it. NOTE: the terms below the recommended terms are not in alphabetical order still.