Changed replaces values for some DCMI terms
In the previous term_versions.csv file, the value in the replaces column indicated what DCMI said the term replaced. However, in the complete history table, we use this column to indicate what this term version replaces with respect to previous TDWG use. For example, in an Executive Decision, dcterms:license was to be used in lieu of dcterms:rights. I've changed the table to reflect this replacement, rather than indicating how http://dublincore.org/usage/terms/history/#rightsHolder-002
 was caused by DCMI to replace http://dublincore.org/usage/terms/history/#rightsHolder-001 .  There are several other examples of replacements caused by TDWG actions and these have replaced the record of DCMI actions. If there was no TDWG-sanctioned replacement, the column was changed to empthy.