Add historical record for deprecated term dcterms:rights
dcterms:rights was originally in Darwin Core, but was replaced by dcterms:license in Executive Decision http://rs.tdwg.org/dwc/terms/history/decisions/#Decision-2014-11-06_17 . It has been added as part of the historical record. It is listed as the "replaces" value for dcterms:license term and needs to be there.