Change link from MD to HTML syntax.
At https://dwc.tdwg.org/, the link to the Maintenance Group in the first paragraph is currently broken - ie, it displays the raw markdown:

    [Darwin Core Maintenance Interest Group](https://www.tdwg.org/standards/dwc/#maintenance%20group)

instead of rendering this as a hyperlink. This may happen because it is embedded within an HTML <p> element and may be fixable by making it into an <a href=> link.